Added Augmentation Process:
Augmentation techniques that we are using are flipping, rotating, brightening and contrast. It works fine and a new folder is added for the augmented data to be placed.

I also added the function for printing the augmented data for checking if the augmentation wa successful. The printing is only for testing and we can either remove it or refine it.